backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
passlib
python-jose
pyjwt
psycopg2-binary
google-auth
